+++ alkos/kernel/include/defines.hpp
@@ -1,30 +1,8 @@
 #ifndef KERNEL_INCLUDE_DEFINES_HPP_
 #define KERNEL_INCLUDE_DEFINES_HPP_
 
-// ------------------------------
-// Attribute macros
-// ------------------------------
-
-/* Prevent the compiler from adding padding to structures */
-#define PACK __attribute__((__packed__))
-
-/* Indicate that the function will never return */
-#define NO_RET __attribute__((noreturn))
-
-/* Force the compiler to always inline the function */
-#define FORCE_INLINE_F inline __attribute__((always_inline))
-
-/* Force the compiler to always inline the lambda */
-#define FORCE_INLINE_L __attribute__((always_inline))
-
-/* Declare a function as a static inline wrapper */
-#define WRAP_CALL static FORCE_INLINE_F
-
-/* Require a function to be inlined for performance reasons */
-#define FAST_CALL static FORCE_INLINE_F
-
-/* Marks a function for a compiler to prevent any optimizations */
-#define NO_OPT __attribute__((optimize("O0")))
+/* defines from libc */
+#include <defines.h>
 
 // ------------------------------------
 // Macro to constexpr conversions
