+++ alkos/libc/include/defines.h
@@ -20,13 +20,29 @@ static constexpr bool kIsKernel = false;
 #endif // __cplusplus
 
 // ------------------------------
-// Lang independent defines
+// Attribute macros
 // ------------------------------
 
-#define INLINE    inline __attribute__((always_inline))
-#define API_CALL  INLINE
-#define WRAP_CALL INLINE
-#define FAST_CALL INLINE
+/* Prevent the compiler from adding padding to structures */
+#define PACK __attribute__((__packed__))
+
+/* Indicate that the function will never return */
+#define NO_RET __attribute__((noreturn))
+
+/* Force the compiler to always inline the function */
+#define FORCE_INLINE_F inline __attribute__((always_inline))
+
+/* Force the compiler to always inline the lambda */
+#define FORCE_INLINE_L __attribute__((always_inline))
+
+/* Declare a function as a static inline wrapper */
+#define WRAP_CALL static FORCE_INLINE_F
+
+/* Require a function to be inlined for performance reasons */
+#define FAST_CALL static FORCE_INLINE_F
+
+/* Marks a function for a compiler to prevent any optimizations */
+#define NO_OPT __attribute__((optimize("O0")))
 
 // ------------------------------
 // Lang specific defines
