+++ alkos/libc/libssp/libssp.cpp
@@ -3,7 +3,7 @@
 // ------------------------------
 
 /* main include */
-#include <libssp.hpp>
+#include <libssp.h>
 
 /* external includes */
 #include <stdint.h>
@@ -36,7 +36,7 @@ volatile uintptr_t __stack_chk_guard = kStackChkGuard;
 
 static void __stack_chk_init_hosted() {}
 
-static __attribute__((noreturn)) void __stack_chk_fail_hosted() {}
+static NO_RET void __stack_chk_fail_hosted() {}
 
 #endif  // __STDC_HOSTED__
 
@@ -54,7 +54,7 @@ static void __stack_chk_init_kernel() {}
 /**
  * @todo Add some debug message about stack in future
  */
-API_CALL static __attribute__((noreturn)) void __stack_chk_fail_kernel()
+WRAP_CALL NO_RET void __stack_chk_fail_kernel()
 {
     KernelPanic("Stack smashing detected");
 }
@@ -78,7 +78,7 @@ void __stack_chk_init()
 #endif
 }
 
-extern "C" __attribute__((noreturn)) void __stack_chk_fail()
+extern "C" NO_RET void __stack_chk_fail()
 {
 #if __STDC_HOSTED__
