+++ alkos/libc/time/gmtime.cpp
@@ -1,0 +1,17 @@
+#include <time.h>
+#include <assert.h>
+
+struct tm *gmtime_s(const time_t *timer, struct tm *result) {
+    assert(false && "Not implemented!");
+    return nullptr;
+}
+
+struct tm *gmtime_r(const time_t *timer, struct tm *result) {
+    assert(false && "Not implemented!");
+    return nullptr;
+}
+
+struct tm *gmtime(const time_t *timer) {
+    assert(false && "Not implemented!");
+    return nullptr;
+}
