+++ alkos/libc/time/localtime.cpp
@@ -1,0 +1,17 @@
+#include <time.h>
+#include <assert.h>
+
+struct tm *localtime_s(const time_t *timer, struct tm *result) {
+    assert(false && "Not implemented!");
+    return nullptr;
+}
+
+struct tm *localtime_r(const time_t *timer, struct tm *result) {
+    assert(false && "Not implemented!");
+    return nullptr;
+}
+
+struct tm *localtime(const time_t *timer) {
+    assert(false && "Not implemented!");
+    return nullptr;
+}
