+++ alkos/libc/time/mktime.cpp
@@ -1,0 +1,7 @@
+#include <time.h>
+#include <assert.h>
+
+time_t mktime(struct tm *time_ptr) {
+    assert(false && "Not implemented!");
+    return time_t{};
+}
