+++ alkos/libc/time/strftime.cpp
@@ -1,0 +1,7 @@
+#include <time.h>
+#include <assert.h>
+
+size_t strftime(char *s, size_t max_size, const char *format, const tm *time_ptr) {
+    assert(false && "Not implemented!");
+    return size_t{};
+}
