+++ alkos/libc/time/time.cpp
@@ -29,33 +29,3 @@ int timespec_getres(struct timespec *ts, int base) {
     TODO_CLOCKS
     return int{};
 }
-
-char *ctime(const time_t *timer) {
-}
-
-errno_t ctime_s(char *buf, rsize_t bufsz, const time_t *timer) {
-}
-
-size_t strftime(char *s, size_t max_size, const char *format, const tm *time_ptr) {
-}
-
-time_t mktime(struct tm *time_ptr) {
-}
-
-struct tm *localtime_s(const time_t *timer, struct tm *result) {
-}
-
-struct tm *localtime_r(const time_t *timer, struct tm *result) {
-}
-
-struct tm *localtime(const time_t *timer) {
-}
-
-struct tm *gmtime_s(const time_t *timer, struct tm *result) {
-}
-
-struct tm *gmtime_r(const time_t *timer, struct tm *result) {
-}
-
-struct tm *gmtime(const time_t *timer) {
-}
